Add source code next to story
