+++ .storybook/addons.js
@@ -1,1 +1,2 @@
 import '@storybook/addon-options/register'
+import '@storybook/addon-storysource/register'
