+++ .storybook/webpack.config.js
@@ -9,6 +9,16 @@ module.exports = {
   module: {
     rules: [
       {
+        test: /\.stories\.tsx?$/,
+        loaders: [
+          {
+            loader: require.resolve('@storybook/source-loader'),
+            options: { parser: 'typescript' }
+          }
+        ],
+        enforce: 'pre'
+      },
+      {
         test: /\.tsx?$/,
         exclude: /node_modules/,
         loader: 'ts-loader'
