+++ package.json
@@ -22,7 +22,7 @@
     "test": "jest",
     "prepublishOnly": "yarn run clean && yarn run build",
     "postpublish": "yarn run deploy-storybook",
-    "dev": "yarn build && start-storybook -p 9009",
+    "dev": "yarn build-standalone && start-storybook -p 9009",
     "build-commonjs": "tsc --outDir commonjs --inlineSourceMap",
     "build-esm": "tsc --module es2015 --outDir esm --inlineSourceMap",
     "build-standalone": "webpack",
@@ -49,6 +49,7 @@
     "@storybook/addon-info": "^5.2.3",
     "@storybook/addon-links": "^5.2.3",
     "@storybook/addon-options": "^3.4.1",
+    "@storybook/addon-storysource": "^5.2.3",
     "@storybook/addons": "^5.2.3",
     "@storybook/react": "^5.2.3",
     "@storybook/storybook-deployer": "^2.3.0",
