+++ stories/category-based.stories.tsx
@@ -1,16 +1,23 @@
 import React from 'react'
+import { groupBy } from 'lodash'
 import { Pane, Heading, SubHeading, Ul, Code, Button } from 'evergreen-ui'
 import { ConsentManagerBuilder } from '../src'
 import DestinationTile from './destination-tile'
+import { storiesOf } from '@storybook/react'
 
 function Section(props) {
   return <Pane is="section" marginBottom={24} {...props} />
 }
 
-export default () => {
+function byCategory(destinations) {
+  return groupBy(destinations, 'category')
+}
+
+const CategoryBased = () => {
   return (
     <Pane maxWidth={1000} margin={30}>
       <ConsentManagerBuilder
+        onError={e => console.error('Error Handling', e)}
         writeKey="mA3XTMcavCUOQo5DL56VIHWcJMsyhAI7"
         otherWriteKeys={['vMRS7xbsjH97Bb2PeKbEKvYDvgMm5T3l']}
       >
@@ -20,24 +27,35 @@ export default () => {
             saveConsent()
           }
 
+          const categories = byCategory(destinations)
+
           return (
             <form onSubmit={handleSubmit}>
               <Section>
                 <Heading>ACME Would like to track you with the following tools:</Heading>
-                <Ul display="flex" flexWrap="wrap">
-                  {destinations.map(d => (
-                    <DestinationTile
-                      key={d.id}
-                      destination={d}
-                      setPreferences={setPreferences}
-                      preferences={preferences}
-                    />
-                  ))}
-                </Ul>
+
+                {Object.keys(categories).map(cat => {
+                  const destinationsForCategory = categories[cat]
+                  return (
+                    <Pane key={cat} marginTop={20}>
+                      <SubHeading>{cat}</SubHeading>
+                      <Ul display="flex" flexWrap="wrap">
+                        {destinationsForCategory.map(d => (
+                          <DestinationTile
+                            key={d.id}
+                            destination={d}
+                            setPreferences={setPreferences}
+                            preferences={preferences}
+                          />
+                        ))}
+                      </Ul>
+                    </Pane>
+                  )
+                })}
               </Section>
 
               <Section>
-                <SubHeading>Preferences</SubHeading>
+                <Heading>Preferences</Heading>
                 <Code>{JSON.stringify(preferences)}</Code>
               </Section>
 
@@ -59,3 +77,5 @@ export default () => {
     </Pane>
   )
 }
+
+storiesOf('ConsentManagerBuilder', module).add(`Category Based`, () => <CategoryBased />)
